Curse Finished

+++ resources/item.py
@@ -1,4 +1,8 @@
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import (jwt_required,
+                                get_jwt_claims,
+                                jwt_optional,
+                                get_jwt_identity,
+                                fresh_jwt_required)
 from flask_restful import Resource, reqparse
 from models.item import ItemModel
 
@@ -15,8 +19,8 @@ class Item(Resource):
             return item.json()
         return {'message': "item not found"}, 404
 
-    @staticmethod
-    def post(name):
+    @fresh_jwt_required
+    def post(self, name):
         if ItemModel.find_by_name(name):
             return {"message": "Item is already exits!"}, 400
         data_in = Item.parser.parse_args()
@@ -27,7 +31,11 @@ class Item(Resource):
             return {"message": "An error happend while inserting"}, 500
         return {"message": "Item added", "item": item.json()}, 200
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'admin privilage needed.'}, 401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -47,6 +55,7 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
         """connection = sqlite3.connect('data.db')
         cursor = connection.cursor()
@@ -59,5 +68,11 @@ class ItemList(Resource):
             items.append({'name': row[1], 'price': row[2]})
         connection.close()
         return {'items': items}, 200"""
-        return {'items': [item.json() for item in ItemModel.find_all()]}
+
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {'items': items}, 200
+        return {'items': [item['name'] for item in items],
+                'message': 'More info available after login.'}, 200
         # return {'items': list(map(lambda x: x.json(), ItemModel.query.all()))}
